Rozjebane ale działa

+++ MiAGK1/MiAGK1/NTypes.h
@@ -188,8 +188,8 @@ struct float4x4
 struct int3
 {
 	int x, y, z;
-	int3(): x(0.0f), y(0.0f), z(0.0f) {}
-	int3(float x, float y, float z) : x(x), y(y), z(z) {}
+	int3(): x(0), y(0), z(0) {}
+	int3(int x, int y, int z) : x(x), y(y), z(z) {}
 
 	void WriteToConsole();
 };
